Updates to the counting code (#94)
**What changes are proposed in this pull request?**
- The counting/tabulation code has been updated to use multivariable
`table()` instead of nested univariable calls to `table()`
- The `N` statistic is now returned for each level of the variable.
- The `ard_categorical(denominator)` argument is much more general now
and accepts a data frame (as before), an integer, one of `c("column",
"row", "cell")`, or a structured data frame.
- The `ard_crosstab()` function has been removed as `ard_categorical()`
easily handles this case now.
- The package now checks for protected column names, eg
`"...ard_xxxx..."`
- Rather than namespacing rlang, the package is now imported.
- The `categorical_variable_summary_fns()` and
`continuous_variable_summary_fns()` functions have been updated to
include an argument that allows users to supplement the default
statistics through this argument.
- `ard_categorical()` will now only return the statistics requested,
e.g. if a user only wants n and p, N will not be returned although it is
required to calculate p.

**Reference GitHub issue associated with pull request.** _e.g., 'closes
#1'_
closes #90 
closes #88


--------------------------------------------------------------------------------

Reviewer Checklist (if item does not apply, mark is as complete)

- [ ] Ensure all package dependencies are installed:
`devtools::install_dev_deps()`
- [ ] PR branch has pulled the most recent updates from master branch:
`usethis::pr_merge_main()`
- [ ] If a bug was fixed, a unit test was added.
- [ ] Run `pkgdown::build_site()`. Check the R console for errors, and
review the rendered website.
- [ ] Code coverage is suitable for any new functions/features:
`devtools::test_coverage()`
- [ ] `usethis::use_spell_check()` runs with no spelling errors in
documentation

When the branch is ready to be merged:
- [ ] Update `NEWS.md` with the changes from this pull request under the
heading "`# cards (development version)`". If there is an issue
associated with the pull request, reference it in parentheses at the end
update (see `NEWS.md` for examples).
- [ ] Increment the version number using `usethis::use_version(which =
"dev")`
- [ ] Run `usethis::use_spell_check()` again
- [ ] Approve Pull Request
- [ ] Merge the PR. Please use "Squash and merge".

---------

Signed-off-by: Daniel Sjoberg <danield.sjoberg@gmail.com>
Co-authored-by: Becca Krouse <14199771+bzkrouse@users.noreply.github.com>