package installtion checking functions (#112)
**What changes are proposed in this pull request?**
* Adding functions that assist in checking whether a Suggested package
is installed. There are `rlang` functions that do this, but these
functions parse the DESCRIPTION file to get the minimum version required
(if present) and ensure both the package is installed AND an appropriate
version is available.

In the future, we will be implementing ARD functions that require other
packages to calculate the statistics. cards will serve as the basis for
an extension package

**Reference GitHub issue associated with pull request.** _e.g., 'closes
#1'_
related #111 


--------------------------------------------------------------------------------

Reviewer Checklist (if item does not apply, mark is as complete)

- [x] Ensure all package dependencies are installed:
`devtools::install_dev_deps()`
- [x] PR branch has pulled the most recent updates from master branch:
`usethis::pr_merge_main()`
- [x] If a bug was fixed, a unit test was added.
- [x] Run `pkgdown::build_site()`. Check the R console for errors, and
review the rendered website.
- [x] Code coverage is suitable for any new functions/features:
`devtools::test_coverage()`
- [x] `usethis::use_spell_check()` runs with no spelling errors in
documentation

When the branch is ready to be merged:
- [x] Update `NEWS.md` with the changes from this pull request under the
heading "`# cards (development version)`". If there is an issue
associated with the pull request, reference it in parentheses at the end
update (see `NEWS.md` for examples).
- [x] Increment the version number using `usethis::use_version(which =
"dev")`
- [x] Run `usethis::use_spell_check()` again
- [x] Approve Pull Request
- [x] Merge the PR. Please use "Squash and merge".

---------

Signed-off-by: Daniel Sjoberg <danield.sjoberg@gmail.com>
Co-authored-by: Becca Krouse <14199771+bzkrouse@users.noreply.github.com>