Check factor variables for `NA` or completely missing levels in `ard_*` functions (#331)
**What changes are proposed in this pull request?**
* Add checks for factors with no levels, or any levels that are `NA`
into `ard_*` functions (#255)

Provide more detail here as needed.

closes #255

Hi @ddsjoberg, please take a look when you get a chance!

I think the only question I have is if we want to add the factor checks
to the top level of the `ard_missing()` function. Currently they are
just called via an internal call to `ard_continuous()` and the checks
aren't triggered if the factor column with the issue is passed as a
variable to `ard_missing()`.



--------------------------------------------------------------------------------

Pre-review Checklist (if item does not apply, mark is as complete)
- [x] **All** GitHub Action workflows pass with a :white_check_mark:
- [x] PR branch has pulled the most recent updates from master branch:
`usethis::pr_merge_main()`
- [x] If a bug was fixed, a unit test was added.
- [x] Code coverage is suitable for any new functions/features
(generally, 100% coverage for new code): `devtools::test_coverage()`
- [x] Request a reviewer

Reviewer Checklist (if item does not apply, mark is as complete)

- [x] If a bug was fixed, a unit test was added.
- [x] Run `pkgdown::build_site()`. Check the R console for errors, and
review the rendered website.
- [x] Code coverage is suitable for any new functions/features:
`devtools::test_coverage()`

When the branch is ready to be merged:
- [x] Update `NEWS.md` with the changes from this pull request under the
heading "`# cards (development version)`". If there is an issue
associated with the pull request, reference it in parentheses at the end
update (see `NEWS.md` for examples).
- [x] **All** GitHub Action workflows pass with a :white_check_mark:
- [x] Approve Pull Request
- [x] Merge the PR. Please use "Squash and merge" or "Rebase and merge".

---------

Co-authored-by: Ines Gimeno Molina <ines.gimeno_molina@novartis.com>
Co-authored-by: Daniel Sjoberg <danield.sjoberg@gmail.com>