Adding `as_cards_fn()` function (#361)
**What changes are proposed in this pull request?**
* Added functions `as_cards_fn()`, `is_cards_fn()`, and
`get_cards_fn_stat_names()`. These functions assist is creating
functions with attributes enumerating the expected results.
* Updated `ard_continuous()` and `ard_complex()` to return full ARDs
when functions passed are created with `as_cards_fn()`: instead of a
single row output, we get a long ARD with rows for each of the expected
statistic names. (#316)

**Reference GitHub issue associated with pull request.** _e.g., 'closes
#<issue number>'_
closes #316

NOTE: We had spoken about including an argument for specifying the
formatting function as well, which is not included. It turns out the
fmt_fn are assigned a very different part of the function, and it would
have required so strange gymnastics to utilize this information. Also,
both `ard_continuous()` and `ard_complex()` have a `fmt_fn` argument
where the formatting functions can be specified.

After this update is live, we can also begin updating a bunch of {cardx}
functions, so when errors occur we continue to get the structure of ARD
we expect without errors as well. Most ARD functions wrap either
`ard_continuous()` and `ard_complex()` internally FYI



--------------------------------------------------------------------------------

Pre-review Checklist (if item does not apply, mark is as complete)
- [ ] **All** GitHub Action workflows pass with a :white_check_mark:
- [ ] PR branch has pulled the most recent updates from master branch:
`usethis::pr_merge_main()`
- [ ] If a bug was fixed, a unit test was added.
- [ ] Code coverage is suitable for any new functions/features
(generally, 100% coverage for new code): `devtools::test_coverage()`
- [ ] Request a reviewer

Reviewer Checklist (if item does not apply, mark is as complete)

- [ ] If a bug was fixed, a unit test was added.
- [ ] Run `pkgdown::build_site()`. Check the R console for errors, and
review the rendered website.
- [ ] Code coverage is suitable for any new functions/features:
`devtools::test_coverage()`

When the branch is ready to be merged:
- [ ] Update `NEWS.md` with the changes from this pull request under the
heading "`# cards (development version)`". If there is an issue
associated with the pull request, reference it in parentheses at the end
update (see `NEWS.md` for examples).
- [ ] **All** GitHub Action workflows pass with a :white_check_mark:
- [ ] Approve Pull Request
- [ ] Merge the PR. Please use "Squash and merge" or "Rebase and merge".