357 stat methods vignette (#365)
**What changes are proposed in this pull request?**

#357 Add an article detailing how a user can implement their own
statistical methods

I'll open this as a draft for now as it would be good just to get a
review on the content initially if you have time @ddsjoberg?

Also a couple of points/questions:
* Currently I've left in the `{cardx}` example, which would mean adding
`{cardx}` to Suggests. I'm happy to remove, just couldn't remember
exactly what we agreed on in the last meeting.
* I haven't been able render due to `as_cards_fn` being new and
installation restrictions on our SCE - if this is important then I
should be able to work around it to render the article
* Added as an article rather than a vignette
* Are there any other parts of the process which change for adding
documentation like this. (If there are then I can also add to the
CONTRIBUTING guidelines)

closes #357 



--------------------------------------------------------------------------------

Pre-review Checklist (if item does not apply, mark is as complete)
- [ ] **All** GitHub Action workflows pass with a :white_check_mark:
- [ ] PR branch has pulled the most recent updates from master branch:
`usethis::pr_merge_main()`
- [ ] If a bug was fixed, a unit test was added.
- [ ] Code coverage is suitable for any new functions/features
(generally, 100% coverage for new code): `devtools::test_coverage()`
- [ ] Request a reviewer

Reviewer Checklist (if item does not apply, mark is as complete)

- [x] If a bug was fixed, a unit test was added.
- [x] Run `pkgdown::build_site()`. Check the R console for errors, and
review the rendered website.
- [x] Code coverage is suitable for any new functions/features:
`devtools::test_coverage()`

When the branch is ready to be merged:
- [x] Update `NEWS.md` with the changes from this pull request under the
heading "`# cards (development version)`". If there is an issue
associated with the pull request, reference it in parentheses at the end
update (see `NEWS.md` for examples).
- [x] **All** GitHub Action workflows pass with a :white_check_mark:
- [x] Approve Pull Request
- [x] Merge the PR. Please use "Squash and merge" or "Rebase and merge".

---------

Signed-off-by: Daniel Sjoberg <danield.sjoberg@gmail.com>
Signed-off-by: Jack Talboys <42234641+jtalboys@users.noreply.github.com>
Co-authored-by: Daniel Sjoberg <danield.sjoberg@gmail.com>
Co-authored-by: Becca Krouse <14199771+bzkrouse@users.noreply.github.com>